add verifytransaction.go
